requirements.txt
openai==1.10.0
python-dotenv
pytest
neo4j
